test: 테스트에서 영화 API 인터셉트 로직을 setupMovieApiMocks 함수로 분리
